+++ python/GafferSceneUI/AttributeEditor.py
@@ -85,10 +85,12 @@ class AttributeEditor( GafferSceneUI.SceneEditor ) :
 			_SearchFilterWidget( searchFilter )
 
 			self.__locationNameColumn = GafferUI.PathListingWidget.defaultNameColumn
+			self.__visibilityColumn = GafferSceneUI.Private.VisibilityColumn( self.settings()["in"], self.settings()["editScope"] )
 			self.__pathListing = GafferUI.PathListingWidget(
 				GafferScene.ScenePath( self.settings()["in"], self.context(), "/", filter = self.__filter ),
 				columns = [
 					self.__locationNameColumn,
+					self.__visibilityColumn,
 				],
 				selectionMode = GafferUI.PathListingWidget.SelectionMode.Cells,
 				displayMode = GafferUI.PathListingWidget.DisplayMode.Tree,
@@ -220,7 +222,7 @@ class AttributeEditor( GafferSceneUI.SceneEditor ) :
 					section = sections.get( currentSection or None, {} )
 					sectionColumns += [ c( self.settings()["in"], self.settings()["editScope"] ) for c in section.values() ]
 
-		self.__pathListing.setColumns( [ self.__locationNameColumn ] + sectionColumns )
+		self.__pathListing.setColumns( [ self.__locationNameColumn, self.__visibilityColumn ] + sectionColumns )
 
 	def __selectedPathsChanged( self, scriptNode ) :